requirements.txt
anyio==4.0.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-telegram-bot==20.5
sniffio==1.3.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
